// src/app/displays/shrTarea_3/interfaces/tarea-user.interface.ts
export interface TareaInterface {
    fecha_Ini: Date;
    fecha_Fin: Date;
    usuario: any;
    razon_Social: any;
    tarea: number;
    tarea_Descripcion: string;
    tarea_Observacion_1: any;
    tarea_Fecha_Ini: Date;
    tarea_Fecha_Fin: Date;
    tipo_Tarea: number;
    tarea_Tipo_Tarea: any;
    estado_Objeto: number;
    tarea_Estado: any;
    tarea_Comentario: any;
    tarea_Comentario_Fecha_Hora: any;
    tarea_Comentario_Estado: any;
    tarea_Comentario_Id: any;
    user_Tarea: any;
    user_Comentario: any;
    user_Invitado: any;
    referencia_Id: any;
    referencia_Descripcion: any;
    tipo_prioridad: number;
    des_Tipo_Prioridad: any;
}